+++ frontend/e-commerce-service-mocked/src/main.ts
@@ -1,5 +1,0 @@
-import { AppModule } from './app/app.config';
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-platformBrowserDynamic().bootstrapModule(AppModule)
-  .catch(err => console.error(err));